+++ federation-js/src/composition/index.ts
@@ -3,4 +3,4 @@ export * from './composeAndValidate';
 export * from './types';
 export { compositionRules } from './rules';
 export { normalizeTypeDefs } from './normalize';
-export { defaultRootOperationNameLookup } from './utils';
+export { defaultRootOperationNameLookup, compositionHasErrors } from './utils';
